rename file with duplicate name
