+++ src/android/android-backend.cpp
@@ -4,7 +4,7 @@
 
 
 #include "android-backend.h"
-#include "android-uvc.h"
+#include "android-uvc-device.h"
 #include "../types.h"
 #include "usb_host/device_watcher.h"
 #include "usb_host/device_enumerator.h"
