+++ src/android/android-uvc-device.h
